+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Changes
 * Minimum required VRChat Avatar SDK version is now 3.9.0
 * Minimum supported Unity version is now 2022.3
+* The `d4rkAvatarOptimizer` component can now be put anywhere in the avatar hierarchy instead of only on the root.
 
 ### Bug Fixes
 * The component now gets properly shown as `d4rk Avatar Optimizer` in the inspector instead of `D 4rk Avatar Optimizer`. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/pull/166)